Modify the challenge viewing as per MVC
